+++ src/browser/header.js
@@ -1,5 +1,0 @@
-!(function(Geotrigger){
-
-if (!Geotrigger || !Geotrigger.Session) {
-  throw new Error("Couldn't find Geotrigger object");
-}